Refactor clearHandler internal to ClearField
